refactor: reorganise table components file structure
