+++ frontend/src/components/dashboard/table/clickableDeleteIcon.js
@@ -11,7 +11,7 @@ import {
 } from '@chakra-ui/react';
 
 import { DeleteIcon } from '@chakra-ui/icons';
-import { deleteUser } from '../../api/api';
+import { deleteUser } from '../../../api/api';
 
 export const ClickableDeleteIcon = ({ userId, setRequestUpdate }) => {
   const { isOpen, onOpen, onClose } = useDisclosure();
